backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
python-dotenv>=1.0
dj-database-url>=2.1
whitenoise>=6.6
django-cors-headers>=4.3
django-storages[google]>=1.14
requests>=2.31
gunicorn>=21.2
